+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r237-release...0fed16a8e10f4440cf5650e882aa378eab08c265)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r238-release...b1f21b9a7db9f8d3e32e7514aa8f8fbf6206e0c0)
 
-- Fix missing indicators
+- Snippet for enabled click cast on non-cell frames (#199)
+- Update default indicators and add English labels (#208)
+
+- Fix an issue if CELL_BORDER_SIZE == 0
+- Fix range check for evokers
+- Fix blacklist shortcut
+- Fix size of bars
+- Fix 0 duration for texture indicators
+- Fix maxValue of bars
+- Fix indicator list again
+- Fix GetUnitFramesForLGF for Spotlights
+- Macro click-castings are now bound by name
+- Refactor Actions using ObjectPool
+- Update appearance of power filter buttons
+- Update debuffs for Nerub-ar Palace (Thanks Reat)
+- Update TWW debuffs
+- Update dispel icon style
 - Update locales